// types/lca.ts
export enum CaseStatus {
  CERTIFIED = 'Certified',
  DENIED = 'Denied',
  WITHDRAWN = 'Withdrawn',
  CERTIFIED_WITHDRAWN = 'Certified - Withdrawn'
}

export enum VisaClass {
  H1B = 'H-1B',
  H1B1_CHILE = 'H-1B1 Chile',
  H1B1_SINGAPORE = 'H-1B1 Singapore',
  E3_AUSTRALIAN = 'E-3 Australian'
}

export enum WageUnitOfPay {
  YEAR = 'Year',
  MONTH = 'Month',
  BI_WEEKLY = 'Bi-Weekly',
  WEEK = 'Week',
  HOUR = 'Hour'
}
// ...
export interface LCAData {
  lcaCaseNumber: string;
  applicationData: ApplicationData;
  jobData: JobData;
  employerData: EmployerData;
  legalData?: LegalData;
}

export interface JobData {
  jobTitle: string;
  socTitle: string;
  baseAnnualSalary: number;
}

export interface ApplicationData {
  caseNumber: string;
  caseStatus: CaseStatus;
  receivedDate: Date;
  decisionDate: Date;
}

export interface EmployerData {
  name: string;
  city: string;
  state: string;
}

export interface LegalData {
  attorneyName: string;
  attorneyFirmName?: string;
}